+++ src/Models/Common.php
@@ -18,8 +18,10 @@ class Common extends NovaPoshta
 
     /**
      * Справочник форм собственности.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a754ff0d-8512-11ec-8ced-005056b2dbe1 Справочник форм собственности
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -27,13 +29,14 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getOwnershipFormsList';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника формы оплаты.
+     * Работает без авторизации.
      *
-     * @deprecated НЕ ДОКУМЕНТИРОВАНО
+     * @since 2022-11-07 НЕ ДОКУМЕНТИРОВАНО
      *
      * @return array
      */
@@ -41,13 +44,14 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getPaymentForms';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника типов контрагентов.
+     * Работает без авторизации.
      *
-     * @deprecated НЕ ДОКУМЕНТИРОВАНО
+     * @since 2022-11-07 НЕ ДОКУМЕНТИРОВАНО
      *
      * @return array
      */
@@ -55,13 +59,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getTypesOfCounterparties';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника технологий доставки.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a6e189f7-8512-11ec-8ced-005056b2dbe1 Получение справочника технологий доставки
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -69,13 +75,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getServiceTypes';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника описаний груза.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a697db47-8512-11ec-8ced-005056b2dbe1 Получение справочника описаний груза
+     * @since 2022-11-07
      *
      * @param  string|null  $find  Запрос поиска
      * @return array
@@ -89,13 +97,15 @@ class Common extends NovaPoshta
             $this->methodProperties['FindByString'] = $find;
         }
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов шин и дисков.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a66fada0-8512-11ec-8ced-005056b2dbe1 Получение справочника видов шин и дисков
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -103,13 +113,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getTiresWheelsList';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов упаковки.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a6492db4-8512-11ec-8ced-005056b2dbe1 Получение справочника видов упаковки
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -122,13 +134,15 @@ class Common extends NovaPoshta
         $this->getHeight();
         $this->getVolumetricWeight();
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов плательщиков обратной доставки.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a6247f2f-8512-11ec-8ced-005056b2dbe1 Получение справочника видов плательщиков обратной доставки
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -136,13 +150,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getTypesOfPayersForRedelivery';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов плательщиков доставки.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a6247f2f-8512-11ec-8ced-005056b2dbe1 Получение справочника видов плательщиков доставки
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -150,13 +166,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getTypesOfPayers';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов паллет.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a5dd575e-8512-11ec-8ced-005056b2dbe1 Получение справочника видов паллет
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -164,13 +182,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getPalletsList';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов обратной доставки груза.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a5b46873-8512-11ec-8ced-005056b2dbe1 Получение справочника видов обратной доставки груза
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -178,13 +198,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getBackwardDeliveryCargoTypes';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов груза.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a5912a1e-8512-11ec-8ced-005056b2dbe1 Получение справочника видов груза
+     * @since 2022-11-07
      *
      * @return array
      */
@@ -192,13 +214,15 @@ class Common extends NovaPoshta
     {
         $this->calledMethod = 'getCargoTypes';
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, false);
     }
 
     /**
      * Получение справочника видов временных интервалов.
+     * Работает без авторизации.
      *
      * @see https://developers.novaposhta.ua/view/model/a55b2c64-8512-11ec-8ced-005056b2dbe1/method/a56d5c1c-8512-11ec-8ced-005056b2dbe1 Получение справочника видов временных интервалов
+     * @since 2022-11-07
      *
      * @param  string  $recipientCityRef  Ref города
      * @param  string|Carbon|null  $dateTime  Дата интервалов