+++ src/Models/Counterparty.php
@@ -18,7 +18,7 @@ class Counterparty extends NovaPoshta
      * Загрузить список контрагентов отправителей / получателей / третье лицо.
      *
      * @see https://developers.novaposhta.ua/view/model/a28f4b04-8512-11ec-8ced-005056b2dbe1/method/a37a06df-8512-11ec-8ced-005056b2dbe1 Загрузить список контрагентов
-     * @since 2022-11-03
+     * @since 2022-11-07
      *
      * @param  string|null  $find  Поиск по названию
      * @return array
@@ -35,14 +35,14 @@ class Counterparty extends NovaPoshta
             $this->methodProperties['FindByString'] = $find;
         }
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, true);
     }
 
     /**
      * Загрузить список контактных лиц Контрагента.
      *
      * @see https://developers.novaposhta.ua/view/model/a28f4b04-8512-11ec-8ced-005056b2dbe1/method/a3575a67-8512-11ec-8ced-005056b2dbe1 Загрузить список контактных лиц Контрагента
-     * @since 2022-11-06
+     * @since 2022-11-07
      *
      * @param  string  $Ref  Ref контрагента
      * @return array
@@ -55,7 +55,7 @@ class Counterparty extends NovaPoshta
 
         $this->methodProperties['Ref'] = $Ref;
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, true);
     }
 
     /**
@@ -64,6 +64,7 @@ class Counterparty extends NovaPoshta
      * @see https://developers.novaposhta.ua/view/model/a28f4b04-8512-11ec-8ced-005056b2dbe1/method/0ae5dd75-8a5f-11ec-8ced-005056b2dbe1 Создать контрагента с типом Физ лицо
      * @see https://developers.novaposhta.ua/view/model/a28f4b04-8512-11ec-8ced-005056b2dbe1/method/bc3c44c7-8a8a-11ec-8ced-005056b2dbe1 Создать контрагента с типом Юр лицо
      * @see https://developers.novaposhta.ua/view/model/a28f4b04-8512-11ec-8ced-005056b2dbe1/method/b0fdf818-8a8e-11ec-8ced-005056b2dbe1 Создать контрагента с типом Третья особа
+     * @since 2022-11-07
      *
      * @param  string  $firstName  Имя
      * @param  string|null  $lastName  Фамилия
@@ -102,14 +103,14 @@ class Counterparty extends NovaPoshta
             $this->methodProperties['Email'] = $email;
         }
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, true);
     }
 
     /**
      * Скачать параметры Контрагента.
      *
      * @see https://developers.novaposhta.ua/view/model/a28f4b04-8512-11ec-8ced-005056b2dbe1/method/a332efbf-8512-11ec-8ced-005056b2dbe1 Скачать параметры Контрагента
-     * @since 2022-11-06
+     * @since 2022-11-07
      *
      * @param  string  $Ref  Ref контрагента
      * @return array
@@ -120,7 +121,7 @@ class Counterparty extends NovaPoshta
 
         $this->methodProperties['Ref'] = $Ref;
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, true);
     }
 
     /**
@@ -139,13 +140,13 @@ class Counterparty extends NovaPoshta
 
         $this->methodProperties['Ref'] = $Ref;
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, true);
     }
 
     /**
      * Получение данных об Контрагенте по номеру телефона (ФИО и прочее).
      *
-     * @since 2022-11-05 НЕ ДОКУМЕНТИРОВАНО
+     * @since 2022-11-07 НЕ ДОКУМЕНТИРОВАНО
      *
      * @param  string|int  $Phone  Телефон
      * @param  string  $LastName  Фамилия (минимум 3 буквы)
@@ -158,6 +159,6 @@ class Counterparty extends NovaPoshta
         $this->methodProperties['Phone'] = $Phone;
         $this->methodProperties['LastName'] = $LastName;
 
-        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties);
+        return $this->getResponse($this->model, $this->calledMethod, $this->methodProperties, true);
     }
 }